64 Bit support

+++ src/game/Pvs.cpp
@@ -362,7 +362,8 @@ pvsStack_t *idPVS::FloodPassagePVS_r(pvsPortal_t *source, const pvsPortal_t *por
 	pvsArea_t *area;
 	pvsStack_t *stack;
 	pvsPassage_t *passage;
-	long *sourceVis, *passageVis, *portalVis, *mightSee, *prevMightSee, more;
+
+	int *sourceVis, *passageVis, *portalVis, *mightSee, *prevMightSee, more;
 
 	area = &pvsAreas[portal->areaNum];
 
@@ -401,16 +402,16 @@ pvsStack_t *idPVS::FloodPassagePVS_r(pvsPortal_t *source, const pvsPortal_t *por
 		source->vis[n >> 3] |= (1 << (n & 7));
 
 		// get pointers to vis data
-		prevMightSee = reinterpret_cast<long *>(prevStack->mightSee);
-		passageVis = reinterpret_cast<long *>(passage->canSee);
-		sourceVis = reinterpret_cast<long *>(source->vis);
-		mightSee = reinterpret_cast<long *>(stack->mightSee);
+		prevMightSee = reinterpret_cast<int *>(prevStack->mightSee);
+		passageVis = reinterpret_cast<int *>(passage->canSee);
+		sourceVis = reinterpret_cast<int *>(source->vis);
+		mightSee = reinterpret_cast<int *>(stack->mightSee);
 
 		more = 0;
 		// use the portal PVS if it has been calculated
 		if (p->done)
 		{
-			portalVis = reinterpret_cast<long *>(p->vis);
+			portalVis = reinterpret_cast<int *>(p->vis);
 			for (j = 0; j < portalVisLongs; j++)
 			{
 				// get new PVS which is decreased by going through this passage
@@ -822,7 +823,7 @@ idPVS::AreaPVSFromPortalPVS
 int idPVS::AreaPVSFromPortalPVS(void) const
 {
 	int i, j, k, areaNum, totalVisibleAreas;
-	long *p1, *p2;
+	int *p1, *p2;
 	byte *pvs, *portalPVS;
 	pvsArea_t *area;
 
@@ -851,8 +852,8 @@ int idPVS::AreaPVSFromPortalPVS(void) const
 		// store the PVS of all portals in this area at the first portal
 		for (j = 1; j < area->numPortals; j++)
 		{
-			p1 = reinterpret_cast<long *>(area->portals[0]->vis);
-			p2 = reinterpret_cast<long *>(area->portals[j]->vis);
+			p1 = reinterpret_cast<int *>(area->portals[0]->vis);
+			p2 = reinterpret_cast<int *>(area->portals[j]->vis);
 			for (k = 0; k < portalVisLongs; k++)
 			{
 				*p1++ |= *p2++;
@@ -911,7 +912,7 @@ void idPVS::Init(void)
 	areaQueue = new int[numAreas];
 
 	areaVisBytes = (((numAreas + 31) & ~31) >> 3);
-	areaVisLongs = areaVisBytes / sizeof(long);
+	areaVisLongs = areaVisBytes / sizeof(int);
 
 	areaPVS = new byte[numAreas * areaVisBytes];
 	memset(areaPVS, 0xFF, numAreas * areaVisBytes);
@@ -919,7 +920,7 @@ void idPVS::Init(void)
 	numPortals = GetPortalCount();
 
 	portalVisBytes = (((numPortals + 31) & ~31) >> 3);
-	portalVisLongs = portalVisBytes / sizeof(long);
+	portalVisLongs = portalVisBytes / sizeof(int);
 
 	for (int i = 0; i < MAX_CURRENT_PVS; i++)
 	{
@@ -1144,7 +1145,7 @@ pvsHandle_t idPVS::SetupCurrentPVS(const int *sourceAreas, const int numSourceAr
 {
 	int i, j;
 	unsigned int h;
-	long *vis, *pvs;
+	int *vis, *pvs;
 	pvsHandle_t handle;
 
 	h = 0;
@@ -1169,8 +1170,8 @@ pvsHandle_t idPVS::SetupCurrentPVS(const int *sourceAreas, const int numSourceAr
 
 			assert(sourceAreas[i] >= 0 && sourceAreas[i] < numAreas);
 
-			vis = reinterpret_cast<long *>(areaPVS + sourceAreas[i] * areaVisBytes);
-			pvs = reinterpret_cast<long *>(currentPVS[handle.i].pvs);
+			vis = reinterpret_cast<int *>(areaPVS + sourceAreas[i] * areaVisBytes);
+			pvs = reinterpret_cast<int *>(currentPVS[handle.i].pvs);
 			for (j = 0; j < areaVisLongs; j++)
 			{
 				*pvs++ |= *vis++;
@@ -1218,7 +1219,7 @@ idPVS::MergeCurrentPVS
 pvsHandle_t idPVS::MergeCurrentPVS(pvsHandle_t pvs1, pvsHandle_t pvs2) const
 {
 	int i;
-	long *pvs1Ptr, *pvs2Ptr, *ptr;
+	int *pvs1Ptr, *pvs2Ptr, *ptr;
 	pvsHandle_t handle;
 
 	if (pvs1.i < 0 || pvs1.i >= MAX_CURRENT_PVS || pvs1.h != currentPVS[pvs1.i].handle.h ||
@@ -1229,9 +1230,9 @@ pvsHandle_t idPVS::MergeCurrentPVS(pvsHandle_t pvs1, pvsHandle_t pvs2) const
 
 	handle = AllocCurrentPVS(pvs1.h ^ pvs2.h);
 
-	ptr = reinterpret_cast<long *>(currentPVS[handle.i].pvs);
-	pvs1Ptr = reinterpret_cast<long *>(currentPVS[pvs1.i].pvs);
-	pvs2Ptr = reinterpret_cast<long *>(currentPVS[pvs2.i].pvs);
+	ptr = reinterpret_cast<int *>(currentPVS[handle.i].pvs);
+	pvs1Ptr = reinterpret_cast<int *>(currentPVS[pvs1.i].pvs);
+	pvs2Ptr = reinterpret_cast<int *>(currentPVS[pvs2.i].pvs);
 
 	for (i = 0; i < areaVisLongs; i++)
 	{
